refactor: move `handle_media_item_input` to .components
